+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Added
 
+- Added a TAP reporter (`kaocha.report/tap`)
 - Added a new `--print-env` flag to the `:kaocha.plugin.alpha/info` plugin,
   which outputs the Clojure and Java version before each run.
+
 
 ## Fixed